==> data_preparation/R_batch3/step2_extract_asReads.R <==

library(asSeq, lib="/nas02/home/w/e/weisun/R/Rlibs/")

# -------------------------------------------------------------------------
# read in the list of the SNP to be excluded
# -------------------------------------------------------------------------

setwd("/lustre/scr/w/e/weisun/TCGA/hetSNP_EA/")
files = list.files(path = ".", pattern="hetSNP_")

sams0 = gsub("hetSNP_", "", files)
sams0 = gsub(".txt", "", sams0, fixed=TRUE)

m1 = match(setdiff(sams0, sams), sams0)
d1 = scan("../R_batch3/submit-jobs.sh", what=character(), sep="\n")
cat(d1[m1], file="../R_batch3/submit-jobs-1.sh", sep="\n")


#for(i in 1:length(files)){

  f1   = files[i]
  sam1 = sams[i]

  cat("\n", sam1, date(), "\n")
  
  input   = sprintf("../bam/%s_sorted_by_name_uniq_filtered.bam", sam1)
  outputTag  = sprintf("../bam/%s_asCounts_hetSNP_EA", sam1)
  snpList = f1
  
  if(! file.exists(f1)){
    stop("snpList file does not exist") 
  }
  
  extractAsReads(input, snpList, outputTag)
#}
